Improve PKCS #11 support in FIPS-compliant mode
Updated the LDAP SDK to improve PKCS #11 support in FIPS-compliant
mode.  When SSLUtil is configured with one or more key managers, and
if at least one of those is a PKCS11KeyManager, then it will try to
use the SunJSSE provider when getting SSLContext instances, even
when running in FIPS-compliant mode.  This is necessary because it
appears that the Bouncy Castle FIPS-compliant BCJSSE provider does
not work well in conjunction with PKCS #11 tokens.