Add output file for generate self-signed cert
Updated the manage-certificates tool to add optional --output-file
and --output-format arguments to the
generate-self-signed-certificate subcommand.  This allows you to
obtain a PEM or DER representation of the certificate in the same
command used to generate it, whereas it was previously necessary to
use a separate command to export the certificate.