Improve DN escaping logic
Improved the logic used to decide which types of optional escaping
should be used when generating the string representation of a DN or
RDN.  If a DN or RDN was created from a string representation, then
that string representation will generally be preserved.  However, if
a DN or RDN is created from its component parts, then the string
representation will be generated, and the LDAP SDK may choose to
escape more characters than are required by RFC 4514.  Previously,
the LDAP SDK always escaped all ASCII control characters and all
non-ASCII characters.  It will now still escape all ASCII control
characters, but it will only escape non-ASCII characters if it does
not believe that they are displayable.  If it believes that a
non-ASCII character is displayable (including Latin characters with
diacritics and letters, digits, spacing, punctuation, and symbols
from languages that use non-Latin characters), that character now be
left unescaped.  The default behavior can be overridden either
programmatically or through a system property.  The logic used to
generate normalized representations of DNs and RDNs has not changed.