Update the default base64 encoding strategy
Updated the default LDIF base64-encoding strategy to fix an issue
that could cause it to be non-standards-compliant.  As a result of
this change, all ASCII control characters and all non-ASCII data
will be base64-encoded.  It is still possible to override the
behavior (in both standards-compliant and non-compliant ways)
programmatically or with a system property.