Improve ldap-diff legacy compatibility
Updated ldap-diff to provide improved compatibility with a legacy
version of the tool (including with tests used to exercise it).  The
tool will now use a default source bind DN of "cn=Directory Manager"
if no value was provided for that argument, and it will also track
entries that do not exist in either server as a separate data point
from entries that exist but are in sync across both servers.  Note
that even if a nonexistent entry is encountered (likely because the
entry was deleted while the tool was running, or because it was
included in a provided DN file), it will not have any effect on
whether the servers are considered in sync.