Fix an in-memory DS search reference issue
Fixed an issue in the in-memory directory server in which it would
only return search result references for smart referral entries
within the scope of the search if those entries matched the search
filter.  It will now correctly return the reference regardless of
whether the smart referral entry matches the filter.