Make it easier to build on Java 17
Updated the LDAP SDK build process so that it's easier to build the
LDAP SDK with a Java 17 JDK.  It already supports running on Java
17, but there were issues that prevented building the LDAP SDK on
Java 17, including the addition of unexpected files in the javaodc
output that were flagged during sanity checking, and an attempt to
use JVM options that are no longer supported when running unit
tests.

Since the LDAP SDK still supports JVMs as old as Java 7, the build
process isn't clean on newer JVMs.  For best compatibility, we still
recommend on building on Java 7 or 8 for the best compatibility.
Official builds are built on Java 7 for best compatibility, but if
you don't need to support Java 7 clients, then Java provides the
best compatibility with all of the components of the build process.

If you're building on Java 9 or newer, you'll need to add
"-Dcheckstyle.enabled=false" to the build command to tell it to skip
code style checking, as the checkstyle library we use for that
purpose isn't compatible with those Java versions.

Similarly, building on newer Java versions will likely report a
large number of compilation warnings, as well as Javadoc warnings
and errors, but the build process should still complete
successfully.  Also, the Jacoco library used for measuring code
coverage for the unit tests isn't compatible with the at least the
newest Java release, so unit tests will be skipped.