Better balance round-robin with unavailable
Fixed an issue with the round-robin server set that could arise if
any of the servers in the set became temporarily unavailable.  In
such cases, the server set could have connected to some of the
remaining available servers with a higher frequency than others
instead of balancing requests evenly across those servers.