Fix a potential manage-account issue
Fixed an issue that could cause the manage-account tool to fail if
it receives a response with an unrecognized password policy state
operation type.  This is most likely to occur when the version of
the manage-account tool does not align with the version of the
Ping Identity Directory Server instance from which the response was
retrieved.