Fix typos in the OID registry
Fixed typos in the OID registry related to the replace certificate
extended operations.  The word "Request" was misspelled as
"Rqeuest", but because the implementation was changed to use the same
OID for both the request and the response (originally the response
was a generic extended result that didn't have an OID or value), the
intended type of "Extended Request" has been replaced with
"Extended Operation".