Start work on version 2 of a logs API
Started working on a new "v2" API for interacting with access log
messages generated by the Ping Identity Directory Server.  This new
API will provide support for additional log formats (including both
the text-based "name=value" format that the server uses by default,
as well as log messages formatted as JSON objects).  It will also
provide better syntax-aware handling for log fields, including
support for syntax-aware redacting and tokenizing fields for
improved privacy (which may optionally operate on components inside
of a field, like redacting the values of certain attributes in a DN
or filter, or tokenizing certain fields in a JSON object).  It also
defines a new AccessLogField data structure that encapsulates the
name and expected syntax for a log field, and it provides classes
with predefined sets of fields for use in text-formatted and
JSON-formatted log messages.