Log-related class updates
Updated the StringLogFieldSyntax class to change its generic type
from String to CharSequence so that it can operate on other types of
string-based values without forcing the caller to convert those
values to strings.

Updated the TextFormattedAccessLogFields and
JSONFormattedAccessLogFields classes to fix some Javadoc typos and
to rename a couple of constants to make their purpose clearer.