Fix an issue with filter matching
Fixed an issue in the Filter.matchesEntry method that could cause it
to throw an exception rather than an appropriate Boolean response
when evaluating an AND or an OR filter in which one of the nested
components is not valid (for example, if the assertion value does
not conform to the syntax for the associated attribute type).  This
could cause problems for other LDAP SDK components that may rely on
filter matching, like the in-memory directory server or the
ldifsearch command-line tool.