Add additional controls for log v2 API
Updated v2 support for parsing text-formatted and JSON-formatted
access log messages to include support for intermediate client
request and response controls and for the operation purpose request
control.  For text-formatted messages, only a string representation
of the controls is available, but JSON-formatted messages provide
details about the contents of those controls.