Allow SSL-prefixed cipher suites on IBM JVMs
Updated the logic used to select the default set of supported cipher
suites to no longer exclude suites starting with "SSL_" by default
on JVMs with a vendor string that includes "IBM".  IBM JVMs appear
to use the "SSL_" prefix for some or all cipher suites, including
those that are not associated with a legacy SSL protocol.  Also,
added a TLSCipherSuiteSelector.setAllowSSLPrefixedSuites method that
can be used to override the default behavior as needed.