Use base64url when generating log tokens
Updated the log field syntax logic used to generate tokens so that
it uses the base64url variant rather than standard base64, and
therefore uses '-' and '_' instead of '+' and '/'.  This is at least
better when tokenizing value components for DNs, where the plus sign
needs to be escaped because it may otherwise indicate the start of
the next RDN component.