Add JSON encodings for controls
Added support for encoding controls to JSON objects, and for
decoding JSON objects as controls.  There is a generic JSON
representation that works for all controls (in which the JSON object
includes just an OID, criticality, and a base64-encoded
representation of the raw value, if there is a value), but most
controls that take a value have a more user-friendly encoding with
the value expressed as a JSON object with fields for each of the
elements of the control value.