Handle embedded JSON-formatted controls
Updated support for JSON-formatted request and response controls to
make it possible to disallow embedded JSON-formatted request and
response controls.  This is the new default behavior, and if the
embedded control is critical, then an attempt to decode the embedded
controls will result in an exception, while if the embedded control
is non-critical, then it will be ignored with a non-fatal error
message.  The JSONFormattedControlDecodeBehavior class has been
updated to make it possible to choose between this behavior and the
former behavior of allowing JSON-formatted request and response
controls to embed other JSON-formatted request and response
controls.