JSON control support for ldapsearch & ldapmodify
Updated the ldapsearch and ldapmodify command-line tools to add
support for JSON-formatted request and response controls.  If the
--useJSONFormattedRequestControls argument is provided, requests
sent to the server will include a JSON-formatted request control
with appropriate encodings for any other request controls to be
included in the request.  Any JSON-formatted response control
received from the server will automatically have its embedded
controls extracted and processed as if they had been returned with
their regular LDAP encodings.