Eliminate a non-ASCII character in a comment
Updated a comment in a unit test source file to eliminate the use of
a non-ASCII character.  Although non-ASCII characters are perfectly
valid in Java source files, they can potentially cause issues with
some tools.

A new unit test has been added to ensure that none of the LDAP SDK
source code (whether in the LDAP SDK itself or the unit tests) will
be allowed to include non-ASCII characters in the future.