Fix JNDI compatibility for controls and extops
Fixed an issue in the JNDI compatibility support for controls,
extended requests, and extended responses.  Although the JNDI
documentation states that the byte array returned by the
getEncodedValue method for these types of objects will include the
BER type and length, that does not actually appear to be the case.
Since the LDAP SDK used the documented behavior for those methods
rather than the apparent actual behavior, conversion of those
elements between the LDAP SDK and JNDI representations would not
yield the expected results.  To address this, the LDAP SDK now uses
the actual behavior that JNDI exhibits rather than the documented
behavior.  The former behavior can be obtained using system
properties in the event that a JNDI implementation actually does
exhibit the documented behavior.