Include resource-config.json in the jar manifest
Updated the LDAP SDK build process so that it includes the
resource-config.json file in the jar file manifest (in a
native-image subdirectory) rather than providing the file outside
of the jar but in the LDAP SDK zip file.  Including the file in the
manifest eliminates the need to explicitly reference it on the
command line using the GraalVM native-image tool.