Add RFC 9371 and link to IANA PEN registry
Updated the set of LDAP-related specifications to include RFC 9371,
which describes the procedures for registering private enterprise
numbers (PENs), which may be used when defining your own object
identifiers for custom schema elements or other purposes.  The OID
registry document has also been updated to provide a link to this
RFC and to the IANA PEN registry.