Remove subversion repository info
Updated the README to remove a reference to checking out the LDAP
SDK source code from a subversion repository rather than from git.
GitHub has announced that they will be discontinuing their support
for subversion access to git repositories, so we will no longer
provide instructions for accessing the code via subversion.

This discontinuation also means that we will stop maintaining a
mirror of the source code in a SourceForge subversion repository
(which was the original location for the open source LDAP SDK
project) because the process we used for mirroring the source code
depended on GitHub's subversion support.