Allow customizing conn not usable result codes
Added an option for customizing the set of result codes that the
LDAP SDK may use to determine whether a connection may no longer be
usable.  Previously, it used a hard-coded set of result codes for
this purpose, and that set is still the default, but it is now
possible to override that with a provided set of result codes.