Add support for using SOCKS proxies
Added a new SOCKSProxySocketFactory class that can be used to allow
establishing LDAP and LDAPS connections through a SOCKSv4 or SOCKSv5
proxy server.  Communication with the SOCKS proxy server itself must
be unencrypted (although communication with the target LDAP server
may optionally be end-to-end encrypted with TLS), and the SOCKS
proxy server must not require authentication.