Add support for HTTP proxy servers
Refactored the support for SOCKS proxy servers to create a new
ProxySocketFactory superclass, and added an HTTPProxySocketFactory
subclass that can be used to tunnel LDAP connections through an
HTTP proxy server.  As with the SOCKSProxySocketFactory, the HTTP
proxy server must not require authentication, and communication with
the proxy server itself will be unencrypted, but the LDAP
communication can optionally be end-to-end encrypted with the target
LDAP server.