Use a better filter for IMDS.getEntry
Updated InMemoryDirectoryServer.getEntry to use a filter of "(&)"
instead of "(objectClass=*)" when searching for the specified entry.
This should be logically equivalent in the vast majority of cases,
but there may be a difference in a corner case in which the server
is configured without schema and you add an entry without the
objectClass attribute.