Add JSONNumber.getValuesAsXXX methods
Updated the JSONNumber class to add getValueAsInteger and
getValueAsLong methods that will return the value of the number as
an Integer or Long, respectively, but only if it can do so
losslessly.  These methods will return null if the value has a
fractional component or if it is outside the range of the associated
data type.

Updated the JSONNumber class to add a getValueAsDouble method that
will return the value of the number as a double.  This method will
never return null, although the value that is returned may have less
precision than the internal BigDecimal value, or it may be converted
to positive or negative infinity if the BigDecimal value is outside
the range that can be represented by a double.