Better handle empty comps in substring filters
Updated the Filter methods for creating substring filters to better
support empty components.  In LDAP filters, filters are transmitted
using a binary encoding, and substring components may be absent but
not empty.  However, because developers creating LDAP-enabled
applications are more likely to be familiar with the string
representations of filters, it is understandable if someone were to
try to provide an empty string instead of null to indicate that the
component is not needed.  The methods for creating substring filters
will now interpret empty components as equivalent to being null,
causing them to be omitted from the resulting encoded LDAP filter.