Add support for the ForgeRock CSN control
Added client-side support for the change sequence number request and
response controls as used in the ForgeRock Directory Server.  The
request control may be included in an add, delete, modify, or modify
DN request to indicate that the server should return a corresponding
response control containing the replication change sequence number
(CSN) that the server has assigned to that operation.