Add Filter.createSubstringAssertion methods
Added createSubstringAssertion methods to the filter class, which
make it easier to construct the string representation of a substring
assertion with a given combination of subInitial, subAny, and/or
subFinal elements.  This can be helpful for creating the assertion
value for an extensible matching filter for use with a substring
matching rule.