Change --helpDebug to --help-debug
Changed the default identifier used to obtain debug help information
from --helpDebug to --help-debug to be more consistent with the
name of the corresponding argument in the Ping Identity Directory
Server.  Either identifier can be used to obtain the help text, but
only --help-debug will be displayed in the tool usage information.