Set FIPS-related props if appropriate in Ping DS
Updated the BouncyCastleFIPSHelper class so that it will set
certain system properties used by the Bouncy Castle FIPS provider if
the LDAP SDK is being used in FIPS-compliant mode and is running as
part of the Ping Identity Directory Server or one of its associated
tools.  This will not have any effect for uses of the LDAP SDK in
non-FIPS-compliant mode, nor for uses in FIPS-compliant mode that
are not part of the Ping Identity Directory Server product.